requirements.txt
gradio
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
langchainhub
gradio
gradio-client
faiss-cpu
openai
pypdf
python-dotenv